Add support for Elixir 1.9+ releases
This commit will also make distillery an optional dependency. This means that you can still create firmware using distillery, but you would need to add {:distillery, "~> 2.0"} to your project's mix file.
